Edit page shows correct data
clicking on rows in the datatable now takes users to the media's edit page. the edit page contains all of the data for whichever show is being viewed. images are technically working right now, but I think they need to be changed so that they are stored either based on a filename or based on the mediaID