jpeg and date validations
Date validation should be good to go. The jpeg validation is having an issue in that it shows invalid for a jpeg upload. The code was correctly identifying whether it was a jpeg or not but had a couple other issues that I fixed and now it doesn't correctly identify a jpeg